fix auth_user update and user in requests
When updating the auth_user, password was
set manually, not using set_password(),
which resulted in garbage data in the database
(passwords are stored encrypted).

Requests referring to request.user where
getting the auth_user id, which can be different
from the user id, for example when there is
an admin account created, which doesn't have
any user data, only auth_user data.